requirements.txt
django==4.1.5
djangorestframework==3.14.0
drf-spectacular==0.25.1
pygments==2.14.0
